:lipstick: CategoryDropdown.jsx 추가 및 디자인 수정
